requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiogram-media-group==0.5.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.14.0
cachetools==4.2.4
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
dnspython==2.6.1
Faker==23.2.1
frozenlist==1.4.1
geographiclib==2.0
geopy==2.4.1
google-api-core==1.34.1
google-auth==2.28.1
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.62.0
googletrans==3.1.0a0
greenlet==3.0.3
grpcio==1.62.0
grpcio-status==1.48.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.12
motor==3.3.2
multidict==6.0.5
polib==1.2.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.6.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-i18n==0.3.9
redis==5.0.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.2.1
yarl==1.9.4
